requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-braces==1.13.0
django-cors-headers==2.5.2
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
lxml
nltk==3.4.1
numpy
oauthlib==3.0.1
pandas
Pillow
# psycopg2==2.7.7
psycopg2
psycopg2-binary
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
soupsieve==1.9
urllib3==1.24.1
webencodings==0.5.1
xlrd==1.2.0
